fix: fix exception handling in C++ native code (#244)
## Description

Fix exception handling in C++ native code so that it results in JS
promise getting rejected with an appropriate message. There are two
fixes here:
1. Runtime type information in the current version of React Native is
disabled because one of its dependencies, which results in C++ not being
able to recognize that `std::runtime_error` is inheriting from
`std::exception` so to extract the exception message we need to match
the type exactly. A fix has been introduced in RN repo, but not yet
present in any release:
https://github.com/facebook/react-native/commit/3132cc88dd46f95898a756456bebeeb6c248f20e.
2. When rejecting a promise the old code was accessing JS runtime in an
unsafe way, this is fixed now.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update (improves or adds clarity to existing
documentation)

### Tested on

- [x] iOS
- [x] Android