feat: add memory pressure to models and allow freeing their memory (#362)
## Description

The garbage collector doesn't see the memory allocated on the heap by
host objects, so we need to notify it of the memory occupied by ET
modules, otherwise it will take a long time before freeing it. We set
the memory pressure to the size of the .pte file of the model (an
underestimate of the true size) and add a host function to release the
module memory.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update (improves or adds clarity to existing
documentation)

### Tested on

- [x] iOS
- [x] Android

### Related issues

#254

### Checklist

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have updated the documentation accordingly
- [x] My changes generate no new warnings