feat: port TextEmbeddings to C++ (#411)
## Description

This PR implements TextEmbeddings Module in C++, following a bigger
effort of porting all the modules to C++. The hook now returns
`Float32Array` instead of `number[]` for perf reasons. This might be
refactored in the future if we want to support models in FP16 or other
types. The PR also gets rid of some dead code from `TokenizerModule`,
`ETModule` and now `TextEmbeddingsModule`.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update (improves or adds clarity to existing
documentation)

### Tested on

- [ ] iOS
- [ ] Android

### Testing instructions

<!-- Provide step-by-step instructions on how to test your changes.
Include setup details if necessary. -->

### Screenshots

<!-- Add screenshots here, if applicable -->

### Related issues

<!-- Link related issues here using #issue-number -->

### Checklist

- [ ] I have performed a self-review of my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have updated the documentation accordingly
- [ ] My changes generate no new warnings

### Additional notes

<!-- Include any additional information, assumptions, or context that
reviewers might need to understand this PR. -->