feat: port LLMs to C++ (#415)
## Description

This PR ports the current LLM functionality to C++, getting rid of the
`Runner` within the `ExecutorchLib` framework. I've also made some
changes to the build system.

`tokenizers-cpp`
- Previously the library was linked in `ExecutorchLib` in XCode via a
build script, now it is completely removed from the frameworks.
- I've prebuiilt static libraries from `tokenizers-cpp` repo, which I
uploaded to `common/ios/libs/tokenizers-cpp`, similarly to the pre-build
ExecuTorch binaries.
- The includes for tokenizers-cpp are now in
`react-native-executorch/third-party/include/tokenizers-cpp/tokenizers_cpp.h`
- Made some changes to the `libs` directory structure, please see the
podspec for reference.
- These headers are then included from the llama runner source code.
- Since `tokenizers-cpp` for Android is pre-built with the `ExecuTorch`
aar, I'm not making any changes here. This will need to be updated once
we bump the ExecuTorch runtime and when we can safely get rid of the
aar/jitpack setup. We can keep tokenizers-cpp as our submodule and then
just reference it in Android's CMake.

`runner`
- The runner source along with headers was moved from `ExecuTorchLib` to
`common/runner`, similarly to `ada` and is compiled on the fly when our
library compiles.
- In the current situation, I think that we will be soon able to get rid
of the ET fork and the submodule.




### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update (improves or adds clarity to existing
documentation)

### Tested on

- [x] iOS
- [x] Android

### Testing instructions

<!-- Provide step-by-step instructions on how to test your changes.
Include setup details if necessary. -->

### Screenshots

<!-- Add screenshots here, if applicable -->

### Related issues

<!-- Link related issues here using #issue-number -->

### Checklist

- [ ] I have performed a self-review of my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have updated the documentation accordingly
- [ ] My changes generate no new warnings

### Additional notes

<!-- Include any additional information, assumptions, or context that
reviewers might need to understand this PR. -->