feat: Resource Fetcher Improvement - pause/resume/cancel (#421)
## Description

Implemented utils functions for ResourceFetcher module - mainly
pause/resume/cancel functions for remote downloads.
Refactored the module to make the code more readable. 
Made progress callback function more meaningful for fetching multiple
resources.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update (improves or adds clarity to existing
documentation)

### Tested on

- [x] iOS
- [ ] Android
### Related issues

Closes Issue #387 

### Checklist

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have updated the documentation accordingly
- [x] My changes generate no new warnings

### Additional notes
- Currently the Example References in docs are incomplete and not
generally true. However they give a rough idea about how it all works,
but they definitely need a fix so I'd appreciate if someone could take a
look at it.
- I just noticed that Issue #387 mentions "Add a way to get information
about size of a file that is downloaded". It is not currently
implemented, but with current logic It can be easily done. I'll add this
as soon as I can.
- Currently pausing/resuming/cancelling is separated between fetch and
fetchMultipleResources. However I just noticed, that with current logic
I can easily mix them together - that way there would not be two
methods: `pauseFetching` and `pauseMultipleFetching` but a single method
`pauseFetching` is going to be sufficient for all cases.
- Although all this imperfections I think this code is in a state where
it is ready for a review.