fix: Fixed yarn prepare failure (#439)
## Description

`yarn prepare` was failing on main and I couldn't install library
locally properly. It works with those changes.
Reviewers, please check if `yarn prepare` works at main for you - maybe
it is a gimmick of my local setup?
Also bumped required version of node as new builder-bob requires it.

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update (improves or adds clarity to existing
documentation)

---------

Co-authored-by: chmjkb <jakubchmura1607@gmail.com>