refactor: change useMemo to useState in hooks. (#509)
## Description

in React [`useMemo`
docs](https://react.dev/reference/react/useMemo#caveats) they explicitly
state, that `useMemo` is an performance optimization and it can throw
away the cached value if it is beneficial. Therefore relying on `useRef`
or `useState` is better in situations where you need to be sure that
object is created exactly once.

Although the `useRef` hook is more "idiomatic", it lacks a built-in lazy
initializer, requiring a more verbose checks.
That is why, `useState` seems like the best option to make sure that
objects are initialized exactly once in a clean way.

### Introducing Breaking Changes

- [ ] Yes
- [x] No

### Type of change

- [ ] Bug fix (change which fixes an issue)
- [ ] New feature (change which adds functionality)
- [ ] Documentation update (improves or adds clarity to existing
documentation)
- [x] Other (chores, tests, code style improvements etc.)

### Tested on

- [x] iOS
- [ ] Android

### Related Issues 

#499 

<!-- Link related issues here using #issue-number -->

### Checklist

- [ ] I have performed a self-review of my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have updated the documentation accordingly
- [ ] My changes generate no new warnings

### Additional notes

<!-- Include any additional information, assumptions, or context that
reviewers might need to understand this PR. -->