build: fix problems with crashes on iPads (#433)
## Description

This PR updates static libraries(.a) files for iOS, so they support iOS
in version 16 and newer, it also solves the problem with crashes on
iPads. Also missing libkernels_portable was added. I suggest to wait
with this PR until all features are rewritten to CPP and then also
remove `s.ios.vendored_frameworks` line from podspec as it won't be
needed anymore, and currently it overwrites linked static libraries

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update (improves or adds clarity to existing
documentation)

### Tested on

- [x] iOS
- [ ] Android

### Testing instructions

<!-- Provide step-by-step instructions on how to test your changes.
Include setup details if necessary. -->

### Screenshots

<!-- Add screenshots here, if applicable -->

### Related issues

<!-- Link related issues here using #issue-number -->

### Checklist

- [x] I have performed a self-review of my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have updated the documentation accordingly
- [x] My changes generate no new warnings

### Additional notes

<!-- Include any additional information, assumptions, or context that
reviewers might need to understand this PR. -->